requirements.txt
networkx>=2.8.0
numpy>=1.21.0
fa2-python>=0.3.5
matplotlib>=3.5.0
scipy>=1.7.0 